Fix two more pre-existing GCC-10/11 build errors
- VkmResourceHandle::isValid() compared uint64_t id against a bare
  -1 int literal, tripping -Werror=sign-compare on GCC.
- [[maybe_unused]] on a class data member under GCC 10/11's
  -std=c++2a hits a known compiler bug ("attribute ignored" under
  -Werror=attributes). Replaced with a genuine frameIndex() getter
  instead of fighting attribute placement across compilers — this
  also makes the field actually useful rather than just silencing
  a warning about it.

+++ include/vkm/renderer/backend/common/renderer_common.h
@@ -46,7 +46,7 @@ namespace vkm
         }
         const bool isValid() const
         {
-            return (id != -1);
+            return (id != (uint64_t)-1);
         }
         const bool isPooledResource() const
         {
